+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - conjur sync version 13.3 - part 3 (continues from 1.0.36v)
 - switch OIDC tests to use @skip with yaml instead of Before hook
+- Failed authentication requests now return no body, only and error code.
+    ONYX-60466
 
 ### Fixed
 - Fixed pubsub scheduler to not hit CPU, removed rake and calling the code itself